+++ cmd/gen/circleci/runner.go
@@ -59,6 +59,7 @@ func (r *runner) run(ctx context.Context, _ *cobra.Command, _ []string) error {
 			RepoName:         r.flag.RepoName,
 			Language:         r.flag.Language,
 			Flavours:         r.flag.Flavours,
+			SkipATS:          r.flag.SkipATS,
 			HasDockerfile:    hasDockerfile,
 			AppCatalog:       r.flag.AppCatalog,
 			AppCatalogTest:   r.flag.AppCatalogTest,